requirements.txt
flask==3.0.2
python-pptx==0.6.23
